pyproject.toml
[project]
name = "plant-disease-classifier"
version = "0.1.0"
description = ""
authors = [
    {name = "Jesse Montgomery",email = "jlmontie@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch (>=2.7.0,<3.0.0)",
    "torchvision (>=0.22.0,<0.23.0)",
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "kaggle (>=1.7.4.5,<2.0.0.0)",
    "notebook (>=7.4.2,<8.0.0)",
    "plotly (>=6.1.0,<7.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]

[tool.poetry]
packages = [{include = "plant_disease_classifier", from = "src"}]


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
